pyproject.toml
[project]
name = "custom-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "tortoise-orm>=0.21.7",
]
